Minor changes to avoid warnings in visual studio 2015

+++ dlib/dnn/cuda_utils.h
@@ -131,8 +131,8 @@ namespace dlib
 
         struct max_jobs
         {
-            max_jobs(size_t n) : num(n) {}
-            size_t num;
+            max_jobs(int n) : num(n) {}
+            int num;
         };
 
         template <typename Kernel, typename... T>
